requirements.txt
-i https://pypi.org/simple
annotated-types==0.5.0 ; python_version >= '3.7'
anyio==3.7.1 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.4 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
efficientnet-pytorch==0.7.1
exceptiongroup==1.1.2 ; python_version < '3.11'
fastapi==0.100.0
filelock==3.12.2 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
mpmath==1.3.0
networkx==3.1 ; python_version >= '3.8'
numpy==1.25.1
pillow==10.0.0
pydantic==2.0.2 ; python_version >= '3.7'
pydantic-core==2.1.2 ; python_version >= '3.7'
requests==2.31.0 ; python_version >= '3.7'
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.27.0 ; python_version >= '3.7'
sympy==1.12 ; python_version >= '3.8'
torch==2.0.1
torchvision==0.15.2
typing-extensions==4.7.1 ; python_version >= '3.7'
urllib3==2.0.3 ; python_version >= '3.7'
uvicorn==0.22.0
